requirements.txt
asgiref==3.4.1
certifi==2021.5.30
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.9.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.8.6
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.1
whitenoise==5.3.0
